PortfolioManagement_Project
Project done for the portfolio management course where we dealt with 6 securities.

Provide a brief description of each of the stock included in your dataset (maximum 2 lines per stock). E.g., sector, company structure, and balance sheet data (use information available online).

Estimate historical mean and variance covariance matrix of the stocks in your dataset with the methods described below. Estimate the variance covariance matrix with a Shrinkage toward Constant Correlation Approach (consider k=0.2). Estimate the vector of averages with an exponential average (lambda=0.005). Comment on the correlations among the stocks observed in your dataset.

Plot the efficient portfolio frontier (in the standard deviation-mean space) with a risk�free asset (Rf=0.01). Comment on the obtained tangent portfolio (or risky asset portfolio).

Plot the efficient portfolio frontier (in the standard deviation-mean space) with the same risky and risk-free assets but imposing a no short-selling constraint. Comment on the differences on the frontier portfolio with 0.5% monthly expected return with and without the short-selling constraint.

Estimate CAPM alpha and beta for your stocks, use the returns of the Nasdaq 100 (you should obtain returns from Nasdaq prices) as proxy for the market returns. Comment on the results, which of the stocks has generated alpha in the previous years? Assume Rf=0.01.

Compute market implicit returns Π in a Black Litterman framework and compute the market portfolio (assuming that the market comprises only your 6 stocks)

Formulate at least two Black Litterman views for the returns of the stocks in your portfolio for January 2023 (for formulating the views you can use historical averages, news and or public balance sheet data). Comment on the Black Litterman portfolio obtained when considering your views